// modules/transfer/modules/transfer-ach-uni/interfaces/ach-uni-account-interface.ts
export interface ACHUniAccount {
  account: string;
  alias: string;
  currency: string;
  product: number;
  subproduct: number;
  enabled: boolean;
  cif: string;
  consortium: string;
  agency: number;
  mask: string;
  name: string;
  status: string;
  availableAmount: number;
  totalAmount: number;
}